revert irrelevant changes; remove unnecessary comments

+++ AndroidApps/BundleTransport/app/src/main/java/net/discdd/bundletransport/utils/ServerMessage.java
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey;
 @Entity(tableName = "ServerMessages")
 public class ServerMessage {
     @PrimaryKey(autoGenerate = false)
-    public int messageId;
+    public long messageId;
     public String date;
     public String message;
     public boolean read;
